Fix word dropdown to allow proper selection
The word dropdown was selecting the second word by default.
The default has been removed so the placeholder text will be
used instead. The check for an empty selection has also
been fixed so this is being properly detected, ensuring that
all the words in the list can be properly chosen.