More work on meanings and reflections.
Adding meanings and reflections now properly updates the list
of meanings and reflections shown on the page for the current
word without triggering a refresh.

Adding a meaning now shows an alert at the top of the page
that automatically disappears. This gives a look and feel
like many other apps, and handles cleaning up the message
automatically.