Finish refactoring words and meanings/reflections pages
Finished with changes (for now) to these pages. The data for
words, meanings, and reflections is now automatically updated
when new items are added, and all options to add values now
work the same, with the button disabled until a value is
entered and then active, with the field cleared and the button
deactivated when the button is clicked.